Step 6

+++ src/Point.java
@@ -3,7 +3,7 @@ public class Point {
     public final double y;
 
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
